# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Subacquirer Payment Integration API",
        "description": "API for integrating with payment subacquirers (PIX and Withdraw transactions)",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user and get access token",
                "description": "Authenticates a user with email and password, returning a Sanctum API token for subsequent requests. **IMPORTANT**: Copy the token from the response and use it in the 'Authorize' button at the top of this page, or include it in the Authorization header as 'Bearer {token}' for protected endpoints.",
                "operationId": "67d6d0fdd54fba132e1b6f55ca254183",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email address",
                                        "type": "string",
                                        "format": "email",
                                        "example": "clientea@example.com"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentication successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Authentication successful"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "description": "Sanctum API token - Copy this value and use it in the Authorization header as 'Bearer {token}' for protected endpoints",
                                                    "type": "string",
                                                    "example": "1|abcdef1234567890abcdef1234567890"
                                                },
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Cliente A"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "clientea@example.com"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Revoke current access token",
                "description": "Revokes the current authenticated user's API token.",
                "operationId": "28d48c13fca984400ac8de99f54ee13a",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pix": {
            "post": {
                "tags": [
                    "PIX Transactions"
                ],
                "summary": "Create a PIX transaction",
                "description": "Creates a new PIX transaction and sends it to the user's configured subacquirer. A webhook will be simulated after 5-10 seconds to update the transaction status. **REQUIRES AUTHENTICATION**: First authenticate via /api/login to obtain a token, then click 'Authorize' at the top of this page and paste your token.",
                "operationId": "e36ce8aafa34728b9bdd3d11b23768d2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "pix_key",
                                    "pix_key_type"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Transaction amount (minimum 0.01)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "pix_key": {
                                        "description": "PIX key (CPF, email, phone, or random key)",
                                        "type": "string",
                                        "example": "12345678900"
                                    },
                                    "pix_key_type": {
                                        "description": "Type of PIX key",
                                        "type": "string",
                                        "enum": [
                                            "cpf",
                                            "email",
                                            "phone",
                                            "random"
                                        ],
                                        "example": "cpf"
                                    },
                                    "description": {
                                        "description": "Optional transaction description",
                                        "type": "string",
                                        "example": "Payment for services",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "PIX transaction created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "PIX transaction created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction_id": {
                                                    "type": "string",
                                                    "example": "PIX-ABCD1234EFGH5678-1699999999"
                                                },
                                                "external_id": {
                                                    "type": "string",
                                                    "example": "ext-123456",
                                                    "nullable": true
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "PENDING"
                                                },
                                                "amount": {
                                                    "type": "string",
                                                    "example": "100.50"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-11-17T21:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error or user/subacquirer issue",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User does not have a subacquirer assigned"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error or subacquirer API error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to process PIX transaction"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "SubadqA API error: Connection timeout"
                                        },
                                        "transaction_id": {
                                            "type": "string",
                                            "example": "PIX-ABCD1234EFGH5678-1699999999"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/withdraw": {
            "post": {
                "tags": [
                    "Withdraw Transactions"
                ],
                "summary": "Create a withdraw transaction",
                "description": "Creates a new withdraw transaction and sends it to the user's configured subacquirer. A webhook will be simulated after 5-10 seconds to update the transaction status. **REQUIRES AUTHENTICATION**: First authenticate via /api/login to obtain a token, then click 'Authorize' at the top of this page and paste your token.",
                "operationId": "516c1444c7ab02721a2c2520630e44f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "bank_code",
                                    "agency",
                                    "account",
                                    "account_type",
                                    "account_holder_name",
                                    "account_holder_document"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Withdraw amount (minimum 0.01)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 500
                                    },
                                    "bank_code": {
                                        "description": "Bank code",
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "agency": {
                                        "description": "Bank agency number",
                                        "type": "string",
                                        "example": "1234"
                                    },
                                    "account": {
                                        "description": "Bank account number",
                                        "type": "string",
                                        "example": "56789-0"
                                    },
                                    "account_type": {
                                        "description": "Account type",
                                        "type": "string",
                                        "enum": [
                                            "checking",
                                            "savings"
                                        ],
                                        "example": "checking"
                                    },
                                    "account_holder_name": {
                                        "description": "Account holder full name",
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "account_holder_document": {
                                        "description": "Account holder CPF/CNPJ",
                                        "type": "string",
                                        "example": "12345678900"
                                    },
                                    "description": {
                                        "description": "Optional transaction description",
                                        "type": "string",
                                        "example": "Withdraw to bank account",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Withdraw transaction created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Withdraw transaction created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction_id": {
                                                    "type": "string",
                                                    "example": "WD-ABCD1234EFGH5678-1699999999"
                                                },
                                                "external_id": {
                                                    "type": "string",
                                                    "example": "ext-456789",
                                                    "nullable": true
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "PENDING"
                                                },
                                                "amount": {
                                                    "type": "string",
                                                    "example": "500.00"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-11-17T21:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error or user/subacquirer issue",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User does not have a subacquirer assigned"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error or subacquirer API error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to process withdraw transaction"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "SubadqA API error: Connection timeout"
                                        },
                                        "transaction_id": {
                                            "type": "string",
                                            "example": "WD-ABCD1234EFGH5678-1699999999"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use the token obtained from /api/login endpoint. Format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "PIX Transactions",
            "description": "PIX Transactions"
        },
        {
            "name": "Withdraw Transactions",
            "description": "Withdraw Transactions"
        }
    ]
}